[generate-fake-db]

+++ packages/shared/src/migrations/1710222133571-createUpcomingVaccinationView.js
@@ -29,7 +29,7 @@ export async function up(query) {
     where s.deleted_at is null
     and s.key = 'vaccine.ageLimit'::text
     union
-    select :ageLimitDefault, 0
+    select :ageLimitDefault::integer, 0
     order by priority desc limit 1
   ),
   vaccine_agelimit as (
